Rename encryption test files so they are found by Surefire
