+++ src/test/java/com/uid2/client/EncryptionV2Tests.java
@@ -10,7 +10,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.Base64;
 
-public class EncryptionTestsV2 {
+public class EncryptionV2Tests {
     @Test
     public void smokeTest() throws Exception {
         UID2Client client = new UID2Client("ep", "ak", CLIENT_SECRET, IdentityScope.UID2);
